+++ tests/unit/specs/src/dapps/Salad/SaladSuccessForm.spec.js
@@ -45,3 +45,4 @@ describe('SuccessForm.vue', () => {
   });
 
 });
+
